requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

transformers #==4.34.1
diffusers #==0.25.0
timm #==1.0.7
xformers #==0.0.22.post7 

accelerate==0.32.1
bitsandbytes==0.42.0

numpy==1.23.5
pandas==2.2.2
scipy==1.14.0
matplotlib==3.9.1
opencv-python==4.10.0.84
pillow==10.2.0
open3d==0.18.0

# instant mesh
pytorch-lightning==2.1.2
omegaconf
einops
xatlas
trimesh
git+https://github.com/NVlabs/nvdiffrast/

# lama
easydict==1.9.0
kornia==0.5.0
albumentations==0.5.2
webdataset

roma
segment_anything
mitsuba==3.5.2 
taichi
plyfile